feat: 기존 검색 컴포넌트 제거 및 최근 검색어 저장 로직 개선
